+++ js/react-image-gallery/package.json
@@ -15,6 +15,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
+    "react-intersection-observer": "^9.5.3",
     "react-router-dom": "^6.21.2"
   },
   "devDependencies": {
